Move intiDb and create test for order-status
