+++ routes/__tests__/authRoute.i.test.js
@@ -8,6 +8,7 @@ import mongoose from "mongoose";
 import { hashPassword } from "../../helpers/authHelper"
 
 import authRoutes from '../authRoute.js'
+import { initDb } from "../../test_utils/utils.js";
 
 const testUser = async () => ({
   name: "Timmy",
@@ -115,7 +116,7 @@ describe("GET /orders", () => {
     ], {
       user
     })
-        
+    
     const res = await request(app).get("/orders")
     
     try {
@@ -128,12 +129,12 @@ describe("GET /orders", () => {
 })
 
 describe("GET /all-orders", () => {
-    it("fails when not admin", async () => {
+  it("fails when not admin", async () => {
     const mongoServer = await MongoMemoryServer.create()
     const connection = mongoose.createConnection(mongoServer.getUri())
     userModel.useConnection(connection)
     orderModel.useConnection(connection)
-
+    
     const user = await new userModel(await testUser()).save()
     
     const app = createExpressTestRoutes([
@@ -167,7 +168,7 @@ describe("GET /all-orders", () => {
     ], {
       user
     })
-        
+    
     const res = await request(app).get("/all-orders")
     
     try {
@@ -179,4 +180,60 @@ describe("GET /all-orders", () => {
   })
 })
 
-// describe("PUT /order-status/:orderId")
+describe("PUT /order-status/:orderId", () => {
+  it("fails when not admin", async () => {
+    const mongoServer = await MongoMemoryServer.create()
+    const connection = mongoose.createConnection(mongoServer.getUri())
+    userModel.useConnection(connection)
+    orderModel.useConnection(connection)
+    const user = await new userModel(await testUser()).save()
+    const { orders } = await initDb(connection)
+    
+    const app = createExpressTestRoutes([
+      ["/", authRoutes]
+    ], {
+      user
+    })
+    
+    const order = orders[0]._id
+    
+    const res = await request(app).put(`/order-status/${order}`).send({status: "cancel"})
+    
+    try {
+      expect(res.status).toBe(401)
+    } finally {
+      connection.close()
+      mongoServer.stop()
+    }
+  })
+  
+  it("succeeds when admin", async () => {
+    const mongoServer = await MongoMemoryServer.create()
+    const connection = mongoose.createConnection(mongoServer.getUri())
+    userModel.useConnection(connection)
+    orderModel.useConnection(connection)
+    
+    const {orders} = await initDb(connection)
+    const user = await new userModel({
+      ...await testUser(),
+      role: 1
+    }).save()
+    
+    const app = createExpressTestRoutes([
+      ["/", authRoutes]
+    ], {
+      user
+    })
+
+    const order = orders[0]
+    
+    const res = await request(app).put(`/order-status/${order._id}`)
+    
+    try {
+      expect(res.status).toBe(200)
+    } finally {
+      connection.close()
+      mongoServer.stop()
+    }
+  })  
+})
